LPFG-570-add-module-blogs: Moved existing methods from learning catalogue into CourseService. Added module get and create methods. Renamed course get, create and list methods.
